+++ lib/sym/app/commands/show_examples.rb
@@ -15,28 +15,36 @@ module Sym
                             echo:    'echo $mykey',
                             result:  '75ngenJpB6zL47/8Wo7Ne6JN1pnOsqNEcIqblItpfg4='.green)
 
-          output << example(comment: 'generate a new password-protected key & save to the default key file',
-                            command: 'sym -gpco ' + Sym.default_key_file,
+          output << example(comment: 'generate a new key with a cached password & save to the default key file',
+                            command: 'sym -gcpqo ' + Sym.default_key_file,
                             echo:    'New Password     : ' + '••••••••••'.green,
                             result:  'Confirm Password : ' + '••••••••••'.green)
 
-          output << example(comment: 'encrypt a plain text string with default key, and save the output to a file',
-                            command: 'sym -es ' + '"secret string"'.bold.yellow + ' -o file.enc',
-                            echo:    'cat file.enc',
-                            result:  'Y09MNDUyczU1S0UvelgrLzV0RTYxZz09CkBDMEw4Q0R0TmpnTm9md1QwNUNy%T013PT0K'.green)
-
-          output << example(comment: 'decrypt a previously encrypted string:',
-                            command: 'sym -ds $(cat file.enc) -k $mykey',
+          output << example(comment: 'encrypt a plain text string with default key file, and immediately decrypt it',
+                            command: 'sym -es ' + '"secret string"'.bold.yellow + ' | sym -d',
                             result:  'secret string'.green)
 
-          output << example(comment: 'encrypt sym.yml and save it to sym.enc:',
-                            command: 'sym -ef sym.yml -o sym.enc -k $mykey')
+          output << example(comment: 'encrypt secrets file using key in the environment, and --negate option:',
+                            command: 'export PRIVATE_KEY="75ngenJpB6zL47/8Wo7Ne6JN1pnOsqNEcIqblItpfg4="',
+                            echo:    'sym -ck PRIVATE_KEY -n secrets.yml',
+                            result:  ''.green)
+
+          output << example(comment: 'encrypt a secrets file using the key in the keychain:',
+                            command: 'sym -gqx keychain.key',
+                            echo:    'sym -ck keychain.key -n secrets.yml',
+                            result:  'secret string'.green)
+
+          output << example(comment: 'encrypt/decrypt sym.yml using the default key file',
+                            command: 'sym -gcq > ' + Sym.default_key_file,
+                            echo:    'sym -n secrets.yml',
+                            result:  'sym -df secrets.yml.enc',
+          )
 
           output << example(comment: 'decrypt an encrypted file and print it to STDOUT:',
-                            command: 'sym -df sym.enc -k $mykey')
+                            command: 'sym -ck production.key -df secrets.yml.enc')
 
-          output << example(comment: 'edit an encrypted file in $EDITOR, ask for key, create file backup',
-                            command: 'sym -tibf secrets.enc',
+          output << example(comment: 'edit an encrypted file in $EDITOR, use default key file, create file backup',
+                            command: 'sym -tbf secrets.enc',
                             result:  '
 Private Key: ••••••••••••••••••••••••••••••••••••••••••••
 Saved encrypted content to sym.enc.
@@ -50,10 +58,10 @@ Diff:
 
           if Sym::App.is_osx?
           output << example(comment: 'generate a new password-encrypted key, save it to your Keychain:',
-                            command: 'sym -gpcx mykey -o ~/.key')
+                            command: 'sym -gpcx staging.key')
 
           output << example(comment: 'use the new key to encrypt a file:',
-                            command: 'sym -e -c -k mykey -f password.txt -o passwords.enc')
+                            command: 'sym -e -c -k staging.key -n etc/passwords.enc')
 
           output << example(comment: 'use the new key to inline-edit the encrypted file:',
                             command: 'sym -k mykey -tf sym.yml.enc')